python/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions===1.20.0
yfinance===0.2.40
pandas===2.2.2
azure-ai-ml===1.17.1
azure-identity===1.17.1
azure-storage-blob===12.20.0
azure-ai-formrecognizer===3.2.0
azure-ai-documentintelligence===1.0.0b2